login and signup things have been done, password reset and email functionality initiated

+++ server/controllers/user-controller.js
@@ -1,15 +1,21 @@
-const User = require("../models/user")
-const { validateEmailAndPasswordForSignup } = require("../middlewares/emailAndPasswordValidation");
+const User = require("../models/user");
+const {
+  validateEmailAndPasswordForSignup,
+} = require("../middlewares/emailAndPasswordValidation");
 const userDetails = require("../models/userDetails");
 const Chat = require("../models/chat");
 const Form = require("../models/form");
 const Url = require("../models/url");
 // const { sendEmail } = require("../utils/sendEmail")
-const { sendEmailWithTemplate } = require("../utils/sendgridEmail")
-const { generateVerificationLink } = require("../utils/generateVerifyLink")
-const { addDataToLogs } = require("./log-controller")
-const { verificationAndBannedCheckForLogin } = require("../middlewares/userMiddleware")
-
+const { sendEmailWithTemplate } = require("../utils/sendgridEmail");
+const { generateVerificationLink } = require("../utils/generateVerifyLink");
+const { addDataToLogs } = require("./log-controller");
+const {
+  verificationAndBannedCheckForLogin,
+} = require("../middlewares/userMiddleware");
+const mongoose = require("mongoose");
+var bcrypt = require("bcryptjs");
+var {sendEmail} = require("../utils/sendEmail")
 
 const login = async (req, res, nxt) => {
   const { user } = req.body;
@@ -27,13 +33,17 @@ const login = async (req, res, nxt) => {
       const token = await existingUser.generateAuthToken();
 
       if (matched) {
-
         var chkAccess = await verificationAndBannedCheckForLogin(existingUser);
 
         if (!chkAccess) {
           await session.abortTransaction(); // Rollback the transaction
           session.endSession();
-          return res.status(401).json({ error: "Your email is not verified, Or you are banned to this platform! Please check your email!" });
+          return res
+            .status(401)
+            .json({
+              error:
+                "Your email is not verified, Or you are banned to this platform! Please check your email!",
+            });
         }
 
         addDataToLogs("User Login", existingUser._id);
@@ -45,17 +55,17 @@ const login = async (req, res, nxt) => {
         return res.status(200).json({
           user: publicProfile,
           token,
-          message: 'User logged in successfully',
+          message: "User logged in successfully",
         });
       } else {
         await session.abortTransaction(); // Rollback the transaction
         session.endSession();
-        return res.status(401).json({ error: 'Password is not correct' });
+        return res.status(401).json({ error: "Password is not correct" });
       }
     } else {
       await session.abortTransaction(); // Rollback the transaction
       session.endSession();
-      return res.status(404).json({ error: 'User not found' });
+      return res.status(404).json({ error: "User not found" });
     }
   } catch (err) {
     await session.abortTransaction(); // Rollback the transaction
@@ -78,50 +88,110 @@ const signup = async (req, res, nxt) => {
     const validationResult = validateEmailAndPasswordForSignup(email, password);
 
     if (validationResult.error) {
-      console.error('Validation error:', validationResult.message);
+      console.error("Validation error:", validationResult.message);
       throw new Error(validationResult.message);
     }
 
     const hashedPassword = await bcrypt.hash(password, 12);
-    var emailVerificationLink = generateVerificationLink()
+    var emailVerificationLink = generateVerificationLink();
     const newUser = new User({
       email,
       firstName,
       lastName,
       password: hashedPassword,
-      emailVerificationLink
+      emailVerificationLink,
     });
     const saveUserPromise = await newUser.save({ session });
 
     const newUserDetails = new userDetails({
-      userID: saveUserPromise._id
+      userID: saveUserPromise._id,
     });
     await newUserDetails.save({ session });
 
     var dynamicTemplateData = {
       firstName,
       lastName,
-      emailVerificationLink
-    }
+      emailVerificationLink,
+    };
     // await sendEmailWithTemplate(process.env["VERIFYEMAILTEMPLATEID"], [email], dynamicTemplateData)
     //   .then(async () => {
-        addDataToLogs("User signUp", saveUserPromise._id);
-        await session.commitTransaction(); // Commit the transaction
-        session.endSession();
-        return res.status(201).json({
-          message: 'Verification Link Sent Successfully, Please check your emailBox for verification!',
-        })
-      // })
-      // .catch((error) => {
-      //   throw error;
-      // });
+    addDataToLogs("User signUp", saveUserPromise._id).then(async () => {
+      await session.commitTransaction(); // Commit the transaction
+      session.endSession();
+      return res.status(201).json({
+        message:
+          "Verification Link Sent Successfully, Please check your emailBox for verification!",
+      });
+    });
+
+    // })
+    // .catch((error) => {
+    //   throw error;
+    // });
   } catch (err) {
     await session.abortTransaction(); // Rollback the transaction
     session.endSession();
     console.error(err);
-    return res.status(500).json({ error: "Something Went Wrong", errorMessage: err.message });
+    return res.status(500).json({ error: err.message });
   }
 };
+
+const forgotPassword = async (req, res, next) => {
+  const { user } = req.body;
+  const { email } = user;
+  const session = await mongoose.startSession();
+
+  try {
+    session.startTransaction();
+
+    var existingUser = await User.findOne({ email: email }).session(session);
+
+    if (existingUser) {
+      var chkAccess = await verificationAndBannedCheckForLogin(existingUser);
+
+      if (!chkAccess) {
+        await session.abortTransaction(); // Rollback the transaction
+        session.endSession();
+        return res
+          .status(401)
+          .json({
+            error:
+              "Your email is not verified, Or you are banned to this platform! Please check your email!",
+          });
+      }
+      const token = await existingUser.generateAuthToken();
+      existingUser.isForgotPasswordInitiated = true;
+      existingUser.forgotPasswordToken = token;
+      var currentDateTime = Date.now();
+      existingUser.forgotPasswordInitiatedDate = currentDateTime;
+      var url = process.env.SERVER_ROOT_URL + "/user/forgot_password_reset?token=" + token + "&email=" + email + "&time=" + currentDateTime;
+      var emailBody = `
+        Hey ${existingUser.firstName}, Here's your link to reset your password: ${url}
+      `
+      await sendEmail("Password Reset Link", [email], emailBody);
+      await existingUser.save();
+      addDataToLogs("User Forgot Password Initiated", existingUser._id);
+      await session.commitTransaction(); // Commit the transaction
+      session.endSession();
+      return res.status(200).json({
+        message: "Reset link sent successfully to youe email 🚀",
+      });
+    } else {
+      await session.abortTransaction(); // Rollback the transaction
+      session.endSession();
+      return res.status(404).json({ error: "User not found" });
+    }
+  } catch (err) {
+    await session.abortTransaction(); // Rollback the transaction
+    session.endSession();
+    console.error(err);
+    return res.status(500).json({ error: err });
+  }
+};
+
+const forgotPasswordReset = async (req, res, next) => {
+
+}
 
 const updateProfile = async (req, res, nxt) => {
   const { user } = req.body;
@@ -130,13 +200,13 @@ const updateProfile = async (req, res, nxt) => {
   session.startTransaction();
 
   const allowedAttributes = [
-    'lastName',
-    'firstName',
-    'email',
-    'bio',
-    'profilePicture',
-    'githubProfile',
-    'githubUsername',
+    "lastName",
+    "firstName",
+    "email",
+    "bio",
+    "profilePicture",
+    "githubProfile",
+    "githubUsername",
   ];
 
   try {
@@ -145,7 +215,7 @@ const updateProfile = async (req, res, nxt) => {
     );
 
     if (extraAttributes.length > 0) {
-      throw new Error(`Invalid attributes: ${extraAttributes.join(', ')}`);
+      throw new Error(`Invalid attributes: ${extraAttributes.join(", ")}`);
     }
 
     await User.update(
@@ -160,50 +230,40 @@ const updateProfile = async (req, res, nxt) => {
           await session.commitTransaction(); // Commit the transaction
           session.endSession();
           return res.status(201).json({
-            message: 'User Updated Successfully!',
+            message: "User Updated Successfully!",
           });
         }
       }
     );
-
   } catch (error) {
     await session.abortTransaction(); // Rollback the transaction
     session.endSession();
     console.error(err);
     return res.status(500).json({ error: "Something Went Wrong" });
   }
-
-}
+};
 
 const deleteProfile = async (req, res, nxt) => {
-
   const session = await mongoose.startSession();
   // starting the mongoose transaction
   session.startTransaction();
 
   try {
+    await User.deleteOne({ _id: req.userId });
 
-    await User.deleteOne(
-      { _id: req.userId }
-    );
+    await userDetails.deleteOne({ userID: req.userId });
 
-    await userDetails.deleteOne(
-      { userID: req.userId }
-    );
+    await Chat.deleteMany({ agent: req.userId });
 
-    await Chat.deleteMany(
-      { agent: req.userId }
-    );
+    await Form.deleteMany({ userID: req.userId });
 
-    await Form.deleteMany(
-      { userID: req.userId }
-    );
+    await Url.deleteMany({ userID: req.userId });
 
-    await Url.deleteMany(
-      { userID: req.userId }
-    );
-
-    await FormDetails.deleteMany({ FormID: { $in: (await Form.find({ userID: req.userId })).map(form => form._id) } });
+    await FormDetails.deleteMany({
+      FormID: {
+        $in: (await Form.find({ userID: req.userId })).map((form) => form._id),
+      },
+    });
 
     await Form.deleteMany({ userID: req.userId });
 
@@ -213,7 +273,7 @@ const deleteProfile = async (req, res, nxt) => {
     session.endSession();
 
     return res.status(201).json({
-      message: 'User Deleted Successfully!',
+      message: "User Deleted Successfully!",
     });
   } catch (error) {
     await session.abortTransaction(); // Rollback the transaction
@@ -221,20 +281,19 @@ const deleteProfile = async (req, res, nxt) => {
     console.error(err);
     return res.status(500).json({ error: "Something Went Wrong" });
   }
-
-}
+};
 
 const getMe = async (req, res, nxt) => {
   try {
     return res.status(201).json({
-      message: 'User Retrived Successfully!',
-      user: await req.rootUser.getPublicProfile()
+      message: "User Retrived Successfully!",
+      user: await req.rootUser.getPublicProfile(),
     });
   } catch (error) {
     console.error(err);
     return res.status(500).json({ error: "Something Went Wrong" });
   }
-}
+};
 
 //--------------------- Admin Controllers-------------------
 
@@ -248,22 +307,20 @@ const getAllUserByAdmin = async (req, res, nxt) => {
     const skip = (page - 1) * limit;
 
     // Query to fetch users with pagination
-    const users = await User.find({})
-      .skip(skip)
-      .limit(limit);
+    const users = await User.find({}).skip(skip).limit(limit);
 
     // Total count of users (you may want to calculate this separately)
     const totalCount = await User.countDocuments();
 
     return res.status(200).json({
-      message: 'User Retrieved Successfully!',
+      message: "User Retrieved Successfully!",
       users,
       totalCount,
     });
   } catch (error) {
     console.error(error);
     return res.status(500).json({
-      error: 'Internal Server Error',
+      error: "Internal Server Error",
     });
   }
 };
@@ -274,7 +331,7 @@ const getUserById = async (req, res) => {
 
     // Check if the provided ID is valid (mongoose.Types.ObjectId)
     if (!mongoose.Types.ObjectId.isValid(userId)) {
-      return res.status(400).json({ error: 'Invalid user ID' });
+      return res.status(400).json({ error: "Invalid user ID" });
     }
 
     // Find the user by ID
@@ -282,12 +339,12 @@ const getUserById = async (req, res) => {
 
     // Check if the user was not found
     if (!user) {
-      return res.status(404).json({ error: 'User not found' });
+      return res.status(404).json({ error: "User not found" });
     }
 
     // User found, send a success response
     return res.status(200).json({
-      message: 'User retrieved successfully',
+      message: "User retrieved successfully",
       user,
     });
   } catch (error) {
@@ -295,7 +352,7 @@ const getUserById = async (req, res) => {
 
     // Handle other errors
     return res.status(500).json({
-      error: 'Internal Server Error',
+      error: "Internal Server Error",
     });
   }
 };
@@ -306,28 +363,23 @@ const deleteUserByAdmin = async (req, res) => {
   session.startTransaction();
 
   try {
+    await User.deleteOne({ _id: req.params.id });
 
-    await User.deleteOne(
-      { _id: req.params.id }
-    );
+    await userDetails.deleteOne({ userID: req.params.id });
 
-    await userDetails.deleteOne(
-      { userID: req.params.id }
-    );
+    await Chat.deleteMany({ agent: req.params.id });
 
-    await Chat.deleteMany(
-      { agent: req.params.id }
-    );
+    await Form.deleteMany({ userID: req.params.id });
 
-    await Form.deleteMany(
-      { userID: req.params.id }
-    );
+    await Url.deleteMany({ userID: req.params.id });
 
-    await Url.deleteMany(
-      { userID: req.params.id }
-    );
-
-    await FormDetails.deleteMany({ FormID: { $in: (await Form.find({ userID: req.params.id })).map(form => form._id) } });
+    await FormDetails.deleteMany({
+      FormID: {
+        $in: (
+          await Form.find({ userID: req.params.id })
+        ).map((form) => form._id),
+      },
+    });
 
     await Form.deleteMany({ userID: req.params.id });
 
@@ -337,7 +389,7 @@ const deleteUserByAdmin = async (req, res) => {
     session.endSession();
 
     return res.status(201).json({
-      message: 'User Deleted Successfully!',
+      message: "User Deleted Successfully!",
     });
   } catch (error) {
     await session.abortTransaction(); // Rollback the transaction
@@ -345,6 +397,17 @@ const deleteUserByAdmin = async (req, res) => {
     console.error(err);
     return res.status(500).json({ error: "Something Went Wrong" });
   }
-}
+};
 
-module.exports = { signup, login, updateProfile, deleteProfile, getMe, getAllUserByAdmin,getUserById, deleteUserByAdmin };
+module.exports = {
+  signup,
+  login,
+  forgotPassword,
+  forgotPasswordReset,
+  updateProfile,
+  deleteProfile,
+  getMe,
+  getAllUserByAdmin,
+  getUserById,
+  deleteUserByAdmin,
+};
